[ext] Condense `dagster-ext` package to one file
